+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\User;
+use Illuminate\Support\Facades\Gate;
+use Spatie\Permission\Models\Role;
 
 class UserController extends Controller
 {
@@ -12,8 +14,14 @@ class UserController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function index()
     {
+        abort_if(Gate::denies('user_index'),403);
         $users = User::all();
 
         return view('users.index',compact('users'));
@@ -26,7 +34,9 @@ class UserController extends Controller
      */
     public function create()
     {
-        return view('users.create');
+        abort_if(Gate::denies('user_create'),403);
+        $roles = Role::all()->pluck('name','id');
+        return view('users.create',compact('roles'));
     }
 
     /**
@@ -37,10 +47,13 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
-        User::create($request->only('name','email')
+
+       $user = User::create($request->only('name','email')
         +[
             'password'=>bcrypt($request->input('password'))
         ]);
+        $roles = $request->input('roles',[]);
+        $user->syncRoles($roles);
         return redirect(route('users.index'));
     }
 
@@ -63,8 +76,11 @@ class UserController extends Controller
      */
     public function edit($id)
     {
+        abort_if(Gate::denies('user_edit'),403);
         $user = User::find($id);
-        return view('users.edit',compact('user'));
+        $roles= role::all()->pluck('name','id');
+        $user->load('roles');
+        return view('users.edit',compact('user','roles'));
 
     }
 
@@ -77,6 +93,7 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
+
         $user = User::findOrFail($id);
         $data = $request->only('name','email');
 
@@ -91,6 +108,8 @@ class UserController extends Controller
         }
 
         $user->update($data);
+        $roles = $request->input('roles',[]);
+        $user->syncRoles($roles);
         return redirect(route('users.index'));
     }
 
@@ -102,9 +121,14 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        abort_if(Gate::denies('user_destroy'),403);
         $user = User::find($id);
-        $user->delete();
 
+        if(auth()->user()->id == $user->id)
+        {
+            return redirect()->route('users.index');
+        }
+        $user->delete();
         return redirect(route('users.index'));
     }
 }
